chore: move rrp logic into new library mod (YARRP)
