+++ .github/workflows/release.yml
@@ -172,6 +172,7 @@ jobs:
         run: |
           echo 'value<<EOF' >> "$GITHUB_OUTPUT"
           echo 'cloth-config(required){modrinth:9s6osm5g}{curseforge:348521}' >> "$GITHUB_OUTPUT"
+          echo 'yarrp(required){modrinth:Z2RtCqwR}{curseforge:1385339}' >> "$GITHUB_OUTPUT"
           if [ "${{ endsWith(matrix.subproject, '-fabric') }}" = "true" ]; then
               # shellcheck disable=SC2129
               echo 'fabric-api(required){modrinth:P7dR8mSH}{curseforge:306612}' >> "$GITHUB_OUTPUT"
