+++ src/main/resources/inventorio.mixins.json
@@ -6,11 +6,9 @@
   "plugin": "de.rubixdev.inventorio.InventorioMixinPlugin",
   "mixins": [
     "ExperienceOrbEntityMixin",
-    "LifecycledResourceManagerImplMixin",
     "PlayerEntityMixin",
     "PlayerInventoryMixin",
     "PlayerManagerMixin",
-    "RegistryEntryListCodecMixin",
     "ServerPlayerEntity_ScreenHandlerSyncHandlerMixin",
     "ServerPlayNetworkHandlerMixin",
     "accessor.CraftingScreenHandlerAccessor",
