+++ versions/1.20.6-neoforge/gradle.properties
@@ -14,6 +14,8 @@
 
 # Common Dependencies
     cloth_version=14.0.139
+    # https://jitpack.io/#RubixDev/YARRP
+    yarrp_version=v0.2.0
     # https://modrinth.com/mod/clumps/versions
     clumps_version=17.0.0.1-neoforge,1.20.6
