+++ versions/1.21.1-common/gradle.properties
@@ -13,6 +13,8 @@
 
 # Common Dependencies
     cloth_version=15.0.140
+    # https://jitpack.io/#RubixDev/YARRP
+    yarrp_version=v0.2.0
     # https://modrinth.com/mod/clumps/versions
     clumps_version=19.0.0.1-fabric,1.21.1
