+++ versions/1.21.1-neoforge/gradle.properties
@@ -14,6 +14,8 @@
 
 # Common Dependencies
     cloth_version=15.0.140
+    # https://jitpack.io/#RubixDev/YARRP
+    yarrp_version=v0.2.0
     # https://modrinth.com/mod/clumps/versions
     clumps_version=19.0.0.1-neoforge,1.21.1
