+++ versions/1.21.1-neoforge/src/main/kotlin/de/rubixdev/inventorio/InventorioNeoForge.kt
@@ -51,6 +51,8 @@ class InventorioNeoForge {
         DeepPocketsBookRecipe.SERIALIZER = serializer
         recipeRegistry.register("deep_pockets_book") { -> serializer }
 
+        InventorioResources.register()
+
         initToolBelt()
         KotlinModLoadingContext.get().getKEventBus().register(InventorioNetworkingNeoForge)
