+++ versions/1.21.1-neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -37,3 +37,10 @@ type = "required"
 versionRange = "[14,)"
 ordering = "NONE"
 side = "CLIENT"
+
+[[dependencies.${mod_id}]]
+modId = "yarrp"
+type = "required"
+versionRange = "[0,)"
+ordering = "NONE"
+side = "BOTH"
